Split the plugin initializtion into two stages: initialization and load.
Added an api module which holds things the plugin can register with to
receive specific event notices as well as become part of the callback
chains for manipulating data.

Added a filestats function to the tools module which wraps os.stat as
well as other members in the api.filestats callback chain.

Added a pyfilenamemtime module that examines file names and if they
match yyyy-mm-dd-HH-MM, then it fixes the mtime on the os.stat tuple.

Refactoring of the pycalendar module so it's a cleaner implementation.  It
also has additional CSS elements (see the file header for more details)
and no longer need the calendarstyle config line in the ini file.