Modified Files: 	libs/api.py libs/pyblosxom.py libs/tools.py Added Files: 	contrib/plugins/logrequest.py
Made logRequest into a callback API call. This allows for statistical
information to be logged via a callback chain. Direct analysis of data can be
done directly based on the current request. This eliminates the need to
actually analyze the log files that may take some plugins, too long to process.