Changed executeHandler and executeChain so they're less confusing (I have no clue what I was thinking when I originally wrote them) and fixed the documentation.
executeHandler serves two purposes.  It allows us to notify all the
plugins of an event in pyblosxom.  It also enables us to dispatch
things to handle to all the plugins--the plugin that handles it
returns api.HANDLED and we stop processing.

executeChain allows us to take data in and pass it through a series
of filters which transform the data.  After the last filter, we
return the newly changed data to the caller.

Thus a CallbackChain is just a list of registered functions and the
way you use those functions is dictated by whether you call
executeChain or executeHandler.