When I fixed tools.Walk last time I was being stupid.  So I fixed it again so that it should show a correct default behavior when passed a None for the pattern.
I also adjusted pycategories so that it shows the number of entries
for each category.  The number of entries for a given category
counts all the subcategories for that category.  For example:

dev/ (26)
  lyntin/ (15)
  pyblosxom/ (2)
  stringbean/ (7)

The category dev has 26 entries:

   15    from lyntin
    2    from pyblosxom
    7    from stringbean
    2    from dev
  ---
   26

Another thing to note is that when you click on a category, you may
not see all the entries for that category--it depends on whether it
spills over the entry max, I think.