Overhauled the code-base to use the new Request object.  The code is not likely all that bug-free at the moment--so be warned.
I also removed the plugins.__init__.load mechanism and replaced it
with the api.prepareChain.  The prepareChain gives plugins a chance
to adjust the data dict and the entry list prior to rendering.

On top of that, I did a bad thing and combined some formatting changes
by converting tabs to spaces in files where I noticed it.