Overhaul with how callbacks are implemented.  Before we had an active-registration callback system.  Plugins would have an initialize function which would get called when they were imported.  In this function they would register with various api callback chains.
Then through PyBlosxom's life cycle we would call either executeHandler,
executeListHandler, or executeTransform on these chains.  The three
different execute* methods differed in how they proceeded through the chain
and what they did with input and output.

The new callback system is a passive-registration system.  When the plugin
is imported, we catalogue all the functions with names that start with cb_
(for callback).  We build the listing of callbacks and their function
lists on the fly--so adding a new callback consists of no steps whatsoever
short of calling run_callback with the new callback and creating a plugin
that has the cb_callbackname function.

We've replaced the execute* methods with an all-powerful run_callback
function in tools.py.  It allows for unique behavior in how it maps the
output of one function to the input of the next function as well as
how to determine whether we're done processing.  These make it far more
powerful than the previous execute* methods since we can essentially
do anything we want.  It also adds the idea of a defaultfunction which
will get called if we've finished running through all the functions in
the callback chain and we haven't satisfied the donefunc conditions.
It might look a little complicated, but it's super duper powerful and
fairly well documented.