Adjusted tools.parse and the Replacer.  I moved the var regexp to a global so that we only create and compile it once.  I also fixed the use of the word "dict" as a variable name.
Fixed pyblosxom so that we sort by mtime using a lambda function rather
than the tools.sortDictBy.  I also fixed it so that if the entry doesn't
have an mtime, it'll float to the top.  I'm pretty sure this nullifies the
requirement for the "mtime" metadata item.

Since sortDictBy isn't used elsewhere in the code-base, I removed it.  It's
kind of a terrible way to sort a list of dicts--better to do it with
cmp and sort.

Added a get method to the entrybase.  It just calls __getattr__--nothing
wild.