Adjusted the variable regexp so that it now handles the following things:
   $foo
   $foo::bar
   $foo::bar::bar2
   $foo::bar("this", "is", "a", "test")

I had to use an eval to grab the arguments (otherwise I would have to parse
them by hand and total ick to that).  It's a slight security issue in the
sense that we're allowing pyblosxom users to eval arbitrary things in their
flavour files.  If the flavour files are locked down (i.e. malicious people
can't edit them), then we should be fine.  Still, it's something to note.