Improved the performance of plugins initialization in the case where pyblosxom never dies and the pyblosxom object was called many times over.
initialize_plugins() was calling catalogue plugins and appending the list of
plugins to the plugins[] list. What happens was that the plugins and callbacks
variables grows larger and larger and in that effect was eating up memory.

The net effect of that is that pyblosxom becomes slower and slower as the list
of plugins grew (though in reality the physical plugins does not).

Short of calling a reload(libs.plugins.__init__) which will refresh the list at
a performance penalty, I opt to return immediately.

Advantages:
  * No more processing of plugins once initialized
  * Improved speed
  * When pyblosxom is running in a continuous state, it does not leak memory
  * Useful for mod_python, fastCGI and twisted
Disadvantages:
  * Once the plugin list is initialized, you cannot drop in another plugin and
    get that to work unless you restart the server

I found out about this problem when doing a twisted pyblosxom