Remove pat and pat_list again :) This time I removed the break as well
The reason why it didn't work previously is due to me forgetting to remove the
break in the previous for loop.. pat requires pattern to be a string, but
pattern is always a re object. so, the result of pat_list is always a [';'].
It didn't make sense at all to do an iteration with one value in a list and the
value is not used at all.

pat and pat_list was a legacy tools.Walk thingy, see:
http://aspn.activestate.com/ASPN/Cookbook/Python/Recipe/52664 for the real
usage of pat and pat_list.