Added some "global" statements to plugin_utils.
Adjusted our handling of the datadir so that we strip any trailing
os.sep on the datadir.  It will always look like "/path/to/blog" now.
In doing this, I discovered we weren't calling Pyblosxom.startup.  I
suspect this was an error, so I injected it into the startup callback
mechanism.  It works fine on my blog, but I'm not entirely sure what
adverse effects it might have elsewhere with all those fine and fancy
things everyone else is doing.