Removed the existing xmlrpc stuff and replaced it with an xmlrpc dispatcher plugin and an xmlrpc blogger plugin.
Added a handle callback.  All plugins that implement the handle callback
get the Pyblosxom request and can do whatever they want with it.  If they
handle the request, they should return a 1.

See contrib/xmlrpc_plugins/xmlrpc.py as an example of a plugin that handles
the request completely--note this involves not going through the rest
of the Pyblosxom lifecycle so you can return whatever HTTP response you
want.

Moved the FieldStorage creation from pyblosxom.cgi to pyblosxom.Pyblosxom.run
after all the plugins have had a chance to handle the request.

xmlrpc plugins now register their methods via a callback which only gets
called when the xmlrpc plugin wants to handle an xmlrpc request.  Implement
the cb_xmlrpc_register callback to register methods.  For an example,
look at contrib/xmlrpc_plugins/xmlrpc_blogger.py.

Retooled the tools.initialize_plugins method to make it so we can call
it multiple times with different plugin lists and plugin dirs.  This
allows xmlrpc plugins to be initialized later on in the cycle since we
probably don't want them showing up in plugininfo since that could alert
someone to the fact that the blogger API is implemented and then maybe
they'll want to try to post to your blog.