Minor documentation changes.
Added a testinstallation mode that allows someone who has just installed
pyblosxom to run "python pyblosxom.cgi" and it'll check their config
file for required items, load all their plugins, and query each plugin
(via verify_installation function) as to whether the plugin is configured
correctly.  This should take a lot of the mystery and annoyance out of
installation and configuration of PyBlosxom and plugins that correctly
implement the "interface".