Changed the way we futz with blog_title.  Before we would change the blog_title variable in the config dict to include the path of the PyBlosxom request.  Some folks really disliked this practice.  On top of that, we're now considering the config dict "read-only".
I changed it so that we create a new "blog_title_with_path" variable in
the data dict which is the blog title with the path.  If you want just
the blog title (with no path ever), use config["blog_title"].  If you
want the blog title with the path, use data["blog_title_with_path"].
This is a backwards-incompatible change.