Reworked command line handling to be more git-like.
pyblosxom-cmd allows for commands implemented in plugins.
Commands are no longer executed by arguments.  For example:

    pyblosxom-cmd --create ...

is now:

    pyblosxom-cmd create ...

It's a pain in the ass to code, but I think it clears up the
command line usage and makes it easier to let plugins do things,
too.  Important things like maintaining the data they're
responsible for.