yeararchives: Don't set flavour to None
In case of urls such as /2003 the new function parse_path_info returns
"None" for flavour and we would set data["flavour"] to None in the
next line (which ends fatal).

This fix basically shuffles that line a few lines further down where
we identified and set the correct flavour properly, so it gets set to
the correct one and can never be None.