Fixed test_when_to_render_comments.
This test case was kind of squirrely, so I ended up reworking bits
of it to match the new behavior for the comments plugin and also
to be more explicit about what it's testing.

I also changed display_comment_default to be either True or False
rather than 1 or 0.

Also, I think some minor cosmetic formatting fixes made it into
the patch.