30. license/copyright documentation overhaul
I looked at how other projects handle license/copyright documentation and
decided to keep a list of authors/contributors in the AUTHORS file, change
license headers in files to point to that files, and add a note about
core plugins many of which have their own copyright/license statements.

I think this achieves better correctness in terms of the documentation,
but also fairly recognizes the work and efforts of all the people involved
without creating a burdensome workflow for maintaining the license/copyright
documentation.