85. Fixes tags plugin initialization problems
The rst_break plugin registers itself with docutils during the start
callback.  That's fine, except that tags wasn't calling the start
callback and thus the rst plugin would bork when trying to parse the
entry since it didn't understand the directives that rst_break was
adding because it hadn't added them, yet.

This fixes things by fixing the initialization routines the tags
plugin uses in buildtags so that it more closes resembles what
PyBlosom does.