Make install guide more readable and comprehensive
Akai pointed out a lot of issues with the install guide and how it flows
which lead to confusion about the various install and deployment options.
These changes are based on his suggested changes with some tweaks.

On the one hand, this makes the documentation harder to maintain--there's
more of it and some bits repeat in various areas.  On the other hand, this
will make it a lot easier for someone new to Pyblosxom to read the Install
guide and have a successful experience.  The latter is important.