Update comments plugin documentation
This factors in Akai's changes, though doesn't use the restructuring
he did.  It does move important information to the top of the documentation
and adds a note about how comments can't be used with static rendering.
It also reworks the Install instructions and makes note about configuration
and adding the flavour templates.