Tweak year/month index settings
This changes the yearnumbers to yearindexes which is a little clearer.
The reason month index settings are monthnames and monthnumbers is that
they're two separate month-based indexes.  There's only one kind of
year indexes, so don't need to name it yearnumbers.

Changed values from 0 and 1 to True and False.

Tweaked the documentation a little.