Fix paginate to work with static rendering.
Previously, paginate didn't work at all with static rendering because it
was using the querystring to denote the page.

This changes that so that if we're rendering a blog statically (i.e.
STATIC is in the data dict), then we denote the page in the url. This
allows us to store all the pages on the file system and everything works
nicely.