Core Python3 fixes
Since has_key is no longer supported change EntryBase to support the in
keyword.
Instead of using BytesIO for the output buffer use StringIO. Since a lot
of code is expecting StringIO this keeps the changes to a minimum. As a
bonus outputting to the terminal does not throw an error either. The
WSGI interface requires a byte string for output so correctly encode it
to bytes before returning.